+++ Efficio/Efficio/Models/Todo+CoreDataClass.swift
@@ -1,0 +1,15 @@
+//
+//  Todo+CoreDataClass.swift
+//  Efficio
+//
+//  Created by k21123kk on 2023/09/26.
+//
+//
+
+import Foundation
+import CoreData
+
+@objc(Todo)
+public class Todo: NSManagedObject {
+
+}
