--- Efficio/Efficio/TodoModel.swift
@@ -1,14 +1,0 @@
-//
-//  TodoModel.swift
-//  Efficio
-//
-//  Created by k21123kk on 2023/09/25.
-//
-
-//import Foundation
-//import CoreData
-//import SwiftUI
-//
-//class TodoModel: ObservableObject {
-//  @Published var date +
-//}
